pyproject.toml
[build-system]
requires = ["flit_core>=3.8,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "apple-hdr-heic"
version = "0.0.3"
authors = [
    { name="John Charankattu", email="john.ch.fr@gmail.com" },
]
description = "A library/tool to decode images in HEIC/HEIF format taken on Apple devices that contain HDR gain map."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Topic :: Multimedia :: Graphics",
  "Topic :: Multimedia :: Graphics :: Graphics Conversion",
]
dependencies = [
  "colour-science>=0.4.5",
  "numpy>=1.24.0",
  "opencv-python-headless>=4.0",
  "openexr>=3.3.2,<4.0",
  "pillow_heif>=0.20.0",
  "PyExifTool>=0.5.6",
]
keywords = ["heic", "heif", "converter", "decoder", "HDR", "HDR Gain Map", "iPhone"]

[project.scripts]
apple-hdr-heic-decoder = "apple_hdr_heic.decoder:main"

[project.urls]
Homepage = "https://github.com/johncf/apple-hdr-heic"

[tool.ruff]
line-length = 120

[tool.mypy]
python_version = "3.10"

[[tool.mypy.overrides]]
module = [
  "exiftool",
  "OpenEXR",
  "pillow_heif",
]
ignore_missing_imports = true

[tool.uv]
# disable cache for easier development
reinstall-package = ["apple-hdr-heic"]

[tool.flit.sdist]
include = [
  "src",
  "pyproject.toml",
  "LICENSE",
  "README.md",
]
exclude = [
  ".gitattributes",
  ".gitignore",
  ".github",
  "noxfile.py",
  "tests",
]
